First attempt at generating a three (four) ported register file using Lattice primitives, with one write port, two read ports, plus an optional read port for debugging purposes.
While here, attach the risc pipeline to the top-level glue logic.  We
don't have a BRAM for code / data yet, so ATM only random noise can be
observed in the instruction pipeline and in the registers, since we
feed the instruction pipeline with a clock tick counter instead with
the proper instruction stream.



git-svn-id: svn+ssh://login.nxlab.fer.hr/fpgasvn/f32c@120 6164e1d0-f123-df11-80f4-0019d13f85ef