Remove unnecessary result-forwarding muxes in the instruction decode stage.
Previously, register file writes were synced on rising clock edge, so
forwarding muxes were necessary to ensure the data to be written into
the register file would be also written to the registers at the boundary
between ID and EX stages.  By clocking register file writes on falling
clock edge, we ensure that the data on read ports of the register file is
up to date before the rising clock edge, so the extra muxes in front of
ID/EX registers become unnecessary.

This change saves 134 LUT4s on Lattice XP2, and is still untested on
Xilinx FPGAs.



git-svn-id: svn+ssh://login.nxlab.fer.hr/fpgasvn/f32c@205 6164e1d0-f123-df11-80f4-0019d13f85ef