Make dhrystone work!
Good news is that the softcore correctly executes the dhrystone test
even at 137.5 MHz, far above the Fmax freqeuncy reported by the
synthesis tool.  The bad is that the results are somewhat on the low
side, only around 62 VAX DMIPS @ 137.5 MHz, or 0.45 DMIPS / MHz, when
compiled with -O3.  According to the PDS dhrystone database, this should
be in range with the first generation Pentium processor @ 75 MHz.



git-svn-id: svn+ssh://login.nxlab.fer.hr/fpgasvn/f32c@418 6164e1d0-f123-df11-80f4-0019d13f85ef