Separate pixel content from compositing metadata. In the metadata is added a pointer to pixel content which is fetched in 1 additional RAM read cycle.
This allows reusing of the same RAM content for different
sprites and sprite animation by only changing the pointer
to data.

As a special case of new compositing2, the metadata
can be organized to point to a contiguous block
of bitmap data. Software written for simple framebuffers
can now work without any change.