Back out 47904a86d6d2402b667f5430a2a5144b9d499d75, which broke ULX2S SRAM on XP2-5E based boards, because among other issues is unnecessarily wasteful on BRAM usage, so the design could not fit in 9 BRAMs.
ULX2S boards with XP2-5E FPGAs are our reference boards which we
MUST continue to support, so please be extra careful not to break
their designs.

Note that this commit only backs out the change to ulx2s_toplevel_sram.vhd
but doesn't touch other files in an attempt to not introduce additional
harm, so the state of our BRAM infrastructure is a complete mess and we
have to revise it carefully.