Clean-slate rewrite of D-cache logic, still disfunctional.
The goal is to optimize for cache hits, and indirectly for Fmax,
by speculatively acknowledging all cache reads in a single cycle,
detect cache misses after the ack has already been propagated to
the CPU pipeline, and in case of misses stall the pipeline to
rectify mispredicted read data in pipelines writeback stage.

Unlike in the original implementation, D-cache BRAMs are dual ported.
Read side is falling-edge triggered, thus allowing a full clock cycle
for read access.  Write side is triggered by clock's rising edges.
No writes are routed to read port, and write port never yields data,
which should make it easy for the tools to map D-cache to device
specific BRAMs, even those available in exotic and low-end devices
such as ICE40.

Nevertheless, currently nothing of this works - it's still unclear
whether new D-cache logic is to blame, or stalling the pipeline
in the writeback stage doesn't work as expected.