Add experimental mutiply-accumulate logic (MI32 only).
On Lattice XP2 this required weird intermediate register placement
in order to appease STA, since the tool assumes the multipler is
always clocked on the positive edge, while in reality it is triggered
on the falling edge, so it was difficult to bring the synthesis and
STA tools in agreement.  Further experimentation is required to
determine how will current pipeline constructs synthesize on other
FPGAs using different synthesis tools.

By default MAC logic is turned off, so unless C_mul_acc is enabled,
this should be a NOP change.