Record branch events even if EX stage is stalled.
Prior to this change, if EX stage was stalled immedeately after a
mispredicted branch (or taken branch in absence of branch predictor),
the information about the taken branch would be lost, which would in
turn lead to the instruction following the branch delay slot to be
propagated to the EX stage, and then executed.

This bug was so far masked by pure luck due to compiler not having
scheduled instructions which could stall the EX stage (MFHI / MFLO,
and more recently MTHI, MTLO, MUL, MADD and MADDU) in the branch
delay slot.