Workaround for a gcc-7.0.1 optimizer bug, which hit malloc_init().
gcc-7.0.1 apparently optimizes out the commented out memory write,
so it has to be forced by casting the pointer to volatile, otherwise
malloc() will never be able to allocate memory, or will work out of
pure luck, which is even worse.

The observed gcc-7.0.1 optimization is obviously invalid and
unwaranted, hence we have to watch closely fresh gcc builds for
similar regressions.