exit(): terminate a .set noreorder block with a .set reorder directive.
The missing .set reorder directive was a causing subseqent instructions
to be arranged in violation of MIPS branch delay slot rules, i.e. it
was possible for another branch or jump to end up in a branch delay
slot, which is both prohibited and undefined.  In turn this was causing
pesky and hard to track bugs when exit(3) would be included in program
sources.