rtl/soc: spi: separate data and status word addresses.
The goal is to make spi.vhd usable on non-f32c platforms which do not
indicate byte-read positions using byte-enable signals, which is a
unique feature (or bug) of f32c.

This will introduce serious churn and bugs throughout the tree, but
is unavoidable if the f32c peripherals are to become gluable to other
platforms.