rtl/cpu: simplify D-cache updating logic.
Remove an optimization which allowed for a previously fetched 32-bit
word to be updated in the cache at byte / half-word granularity.  This
offered a slight increase in CoreMark score but required a mux in front
D-cache write port plus a bit of extra logic, which is now removed.

The net savings is around 100 LUTs with ECP5 synthesis with Diamond 3.9.

The net loss is CoreMark/MHz score drop from 2.917 to 2.903.

DMIPS/MHz score, as well as and block read and write throughput, remain
unaffected.