src/bench/dhry: really prevent any function inlining as required per Dhrystone rules.
Apparently earlier gcc versions did not honor -fno-inline completely,
so the barrier to inling was emulated using a mix of other options, which
however did not completely suppress all function inlining.

An expected consequence is a noticable reduction in DMIPS/MHz scores,
which still has to be more broadly evaluated.