src: revert bbfb525d20e8af9b91170cfa4dcfebf3d21bffbd
Import of clz() / ctz() variants from FreeBSD was hastedly commited
without proper testing.  Not only we already have clzsi2() in
floating point emulation, but the import of clz() and friends created
circular defines with ffs() and friends, which resulted in infinite
recursions and brutal crashes.

clz() / ctz() need to be revised with proper testing.