src/lib: redefine fb_set_mode()
We now support arbitrary video modes defined by the modeline structure
supplied as the first argument to fb_set_mode(), provided that the
attached display will accept it.

For convenience, four video modes are predefined in dev/fb.h: 720p60,
720p50, 1080i60 and 1080i50.  Most if not all of those modes will be
supported, and at least one must be supported, by any of the 16:9
monitors or digital TV-sets claiming to be HD, FHD or better.

The second argument to fb_set_mode() is the BPP depth, and an optional
pixel doubling flag, all of which are also defined in dev/fb.h

So far, the library supports only 8 and 16 BPP.