rtl/soc: vga2dvid: add an extra register before ODDR
Here and there, the timing report from Lattice Diamond complaints that
the path between DVI output shift registers and ODDRs in IOB would
become overly stretched.  Inserting registers on that path gives the
tool more freedom to place DVI logic further away from the IOBs and
reliably resolves the issue.

The cost is essentially zero as FFs are abundant, and the introduced
1/5 pixel clock delay certainly won't be visible on the display...