rtl/soc: introduce a counter of short breaks on a sio line
Detects a sequence of short break pulses (cca. 50 ms) observed on
the input, typically an asynchronous serial line from a controlling
terminal.  The output "sel" will report zero for a single pulse, one for
two consecutive pulses, and so on.  The output may be used as a
selector for a mux, for example.